add reactive IT with Spring properties
