+++ integration-tests/common-basic/src/main/resources/application.yml
