+++ integration-tests/oauth2-spring/src/test/resources/testRunner.yml
@@ -18,7 +18,7 @@ implementation: com.okta.spring.tests.oauth2.spring.SpringApplicationUnderTest
 scenarios:
 
   code-flow-remote-validation:
-    command: com.okta.spring.tests.common.code.BasicRedirectCodeFlowApplication
+    command: com.okta.spring.tests.common.basic.code.BasicRedirectCodeFlowApplication
     args:
       - --server.port=${applicationPort}
       - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
@@ -36,7 +36,7 @@ scenarios:
     protectedPath: /oauth2/authorization/okta
 
   implicit-flow-local-validation:
-    command: com.okta.spring.tests.common.jwt.BasicJwtResourceServerApplication
+    command: com.okta.spring.tests.common.basic.jwt.BasicJwtResourceServerApplication
     args:
     - --server.port=${applicationPort}
     - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
@@ -45,7 +45,7 @@ scenarios:
     - --server.session.trackingModes=cookie
 
   implicit-flow-remote-validation:
-    command: com.okta.spring.tests.common.opaquetoken.BasicOpaqueTokenResourceServerApplication
+    command: com.okta.spring.tests.common.basic.opaquetoken.BasicOpaqueTokenResourceServerApplication
     args:
       - --server.port=${applicationPort}
       - --spring.security.oauth2.resourceserver.jwt.issuer-uri=https://localhost:${mockHttpsPort}/oauth2/default
