+++ pom.xml
@@ -42,10 +42,12 @@
         <module>oauth2</module>
         <module>sdk</module>
         <module>okta-spring-boot-starter</module>
-        <module>integration-tests/common</module>
+        <module>integration-tests/common-basic</module>
+        <module>integration-tests/common-reactive</module>
         <module>integration-tests/oauth2</module>
         <module>integration-tests/oauth2-reactive</module>
         <module>integration-tests/oauth2-spring</module>
+        <module>integration-tests/oauth2-spring-reactive</module>
         <module>examples</module>
         <module>coverage</module>
     </modules>
@@ -94,6 +96,16 @@
             <dependency>
                 <groupId>com.okta.spring</groupId>
                 <artifactId>okta-spring-boot-integration-tests-common</artifactId>
+                <version>2.0.2-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-common-basic</artifactId>
+                <version>2.0.2-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-common-reactive</artifactId>
                 <version>2.0.2-SNAPSHOT</version>
             </dependency>
